Implement the commands and their wrappers to send a fragment to a remote site get a fragment from a remote site.
The internal functions (get/put) from remote.mx are used to implement both functions.
The function print is only for debugging.

All the connection management, which is now done by remote.mx, was removed from the bpm.
The idea is to re-use as much as possible of the remote code.

Note: The code was compiled with debug and optimized mode. The mtest for sql was also ran.