propagated Martin's recent changes of the 64-bit OIDs-specific output to the general (32-bit OID) output: `cvs diff -r1.{5,7} src/mal/Tests/tst274.stable.out.oid64 | patch src/mal/Tests/tst274.stable.out`
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
2008/08/09 - mlkersten: src/mal/Tests/tst274.stable.out.oid64,1.6
The access modes of the kernel BAT update operations were not aligned.
This caused the SQL front-end to inject setWriteMode calls before each
append, and possibly trimming it again with an optimizer.

The BAT insert, append, delete, and replace functions now always return
the first argument, which my be a copy with the proper access properties.
This elicits the flow. The side effectness can be checked with the
isUpdateFunction.

The majority of files involve the now visible temporary in the output.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
2008/08/12 - mlkersten: src/mal/Tests/tst274.stable.out.oid64,1.7
A pass over the generation of MAL listings. The code has been condensed
and some anomalies removed. A cleanup that does not improve functionality
except for ease of access to information during debugging.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~