fixing the leak reported in [ 2540169 ] Memory leak and crash with UNION ALL https://sourceforge.net/tracker/index.php?func=detail&aid=2540169&group_id=56967&atid=482468
by making sure that we use the BAT returned by bat.append since
~~~~~~~
2008/08/09 - mlkersten: MonetDB5/src/modules/kernel/bat5.mx,1.37
The access modes of the kernel BAT update operations were not aligned.
This caused the SQL front-end to inject setWriteMode calls before each
append, and possibly trimming it again with an optimizer.

The BAT insert, append, delete, and replace functions now always return
the first argument, which my be a copy with the proper access properties.
This elicits the flow. The side effectness can be checked with the
isUpdateFunction.

The majority of files involve the now visible temporary in the output.
~~~~~~~