Fix a potential deadlock.
Suppose you have a largish file with SQL queries and that those
queries are sent to the server in 8K (or so) chunks (as is done in the
doFile function).  Also suppose that many of those SQL queries produce
a significant number of rows.  What could happen in this situation is
that the first SQL query to produce lots of output is executed.  Only
the first 100 or 10000 (depending on the age of your mclient) rows are
produced.  The server then continues on with the next query and
produces its results (or the first 100/10000 rows).  mclient consumes
the rows of the first query and when at the end of the 100/10000 rows
it needs to ask the server for more results for the query.  This means
that mclient has to read all results until it gets to a server prompt
so that it can send the appropriate X command (X-export).  However,
chances are that the last part of the query set that mclient had sent
ends in the middle of a query.  This means that the server at this
stage needs to read more of a query and cannot cope with an X
command.  mclient at the same time is somewhere deep in the Mapi
library reading output from the server and it cannot send the rest of
the queries that the server craves.

The solution is to not limit the output of the server to only the
first 100/10000 rows of each query but to let the server send it all
so that the client doesn't have to send X-export commands.