make it possible to set - gdk_mem_maxsize - gdk_vm_maxsize by hand.
gdk_mem_maxsize is the amount of phyiscal memory MonetDB is allowed to
use, and thus is an important parameter that should be set by the user
(by default, MonetDB thinks it is entitled to 80% of your RAM).

gdk_vm_maxsize is no issue on 64-bits, but user might want to increase it
to 2.5GB on systems with a 3GB user space, or to 3.5GB when running
on a 64-bits system in 32-bits mode (default is 1.5GB on 32-bits).