Undoing my recent checking (see quote below):
Having non-power-of-2 wide atom in BUN heaps aligned on power-of-2
boundaries was not a good idea, since this prevents up from accessing BUN
heaps as arrays of the respective type.

Alternative solutions are
1) align also non-power-of-2 wide atom on their actual width, and thus loose
   the ability to calculate addresses from indices and v.v. by simple bit
   shifting;
2) require all atom to be power-of-2 wide, i.e., pad them where necessary to
   meet this requirement.

One of them will follow in my next checkin...

-------------------------------------------------------------------
2009/08/12 - stmane: MonetDB4/src/modules/contrib/bat_cast.mx,1.8.6.1
        MonetDB4/src/modules/contrib/bat_mmath.mx,1.10.6.1
        MonetDB4/src/modules/plain/bat.mx,1.35.2.1
        MonetDB4/src/modules/plain/xtables.mx,1.18.4.1
        MonetDB4/src/monet/monet_interpreter.mx,1.25.4.1
        MonetDB4/src/monet/monet_multiplex.mx,1.16.4.1
        MonetDB5/src/modules/kernel/bat5.mx,1.56.2.1
        MonetDB5/src/modules/kernel/group.mx,1.137.2.1
        MonetDB5/src/modules/mal/bpm.mx,1.142.4.1
        MonetDB5/src/modules/mal/mkey.mx,1.19.2.1
        MonetDB/src/gdk/gdk.mx,1.287.2.1 MonetDB/src/gdk/gdk_align.mx,1.103.2.1
        MonetDB/src/gdk/gdk_bat.mx,1.222.2.1
        MonetDB/src/gdk/gdk_batop.mx,1.173.2.1
        MonetDB/src/gdk/gdk_relop.mx,1.163.2.1
        MonetDB/src/gdk/gdk_scanselect.mx,1.58.2.1
        MonetDB/src/gdk/gdk_setop.mx,1.71.2.1

fixed bug that surfaced after
"
2009/07/25 - nielsnes: MonetDB5/src/modules/kernel/group.mx,1.134
return :wrd on counts

added optimizations when the input is ordered (derive and refine cases)
"
with many group-gy tests on 64-bit systems with 32-bit OIDs:

The above checkin made the mapentry_t struct in MonetDB5/src/modules/kernel/group.mx
grow from 4*4 = 16 bytes to 3*4+8=20 bytes (or better 24-bytes with 8-byte alignment)
on 64-bit systems with 32-bit OIDs (while staying a power-of-2 size on both
64-bit systems with 64-bit OIDs (4*8=32 bytes) and 32-bit systems (4*4=16 bytes)).

This change revealed that MonetDB (GDK) as well as MonetDB4 and MonetDB5 did not
correctly handle non power-of-2 wide atom in BATs in numerous places.
In detail, for simple and fast address calculation via bit-shifts, MonetDB stores
values of fixed-width atoms in BATs (head & tail) on power-of-2 aligned addresses,
even is the atom width is not a power-of-2. In numerous places, the code incorrectly
used the actual atom width instead of the power-of-2 alignement to calculate addresse,
resulting in incorrect behavior.

Fixes:
- fixed implementation of function ATOMelmshift() in MonetDB/src/gdk/gdk_bat.mx
- replaced actual atom width (Hsize(b), Tsize(b), H->width, T->width)
  by power-of-2 aligned atom width (headsize(b,1), tailsize(b,1))
  whereever appropriate

For convenience, these changes are wrapped in CVS tags
"H-T_size" & "head-tail_size".
-------------------------------------------------------------------