Changed the implementation of the offset heap for type string.
Before, the offset heap would contain values of type var_t (typically
equal in size to oid), so if there are lots of entries but not that
many different strings, a lot of space was wasted, since only the
lower few bytes of the values in the offset heap would be used, the
rest being zero.

In this version, the offset heap has a dynamic width.  It starts off
as a single byte value, and when the offsets become too large to fit
in a single byte, the heap is upgraded to a short, then an int, and
finally (if the system uses 64 bit oids) to lng.

There are tags to mark this change:
before_variable_string_offset and after_variable_string_offset.

Only MonetDB* and pathfinder needed to be changes, so the tags are
only added there.