reverting " 2010/02/22 - mlkersten: MonetDB5/src/optimizer/opt_joinpath.mx,1.62.2.9 A join can be injected into a leftjoinpath structure. This may lead to a more efficient evaluation. "
Extensive testing with TPCH did not show any noticeable performance
differences with and without this feature.
However, with SSB performance drops by up to a factor 3 for single
queries and a factor 1.5 to 2 overall with this feature.
The reason is most probably that "restricting" a join to a leftjoin
implementation, i.e., an implementation that is bound to produce the
join result in the order of the left input and hence is not allowed
to swap the inputs to exploit their properties for a more efficient
join evaluation, comes with a penalty that exceeds the potential
benefits of the (left)joinpath optimization.