Make that views on bats don't cause the parent bat to remain loaded.
When a bat has a view, the physical reference of the parent bat is
incremented, causing it to stay in core while the view exists.
However, the parent bat also remains in core when the view has a
physical reference of 0 (but with a logical reference > 0 so that the
view does still exist).

This patch changes that.

When a view's physical reference count goes to zero, the paren't
reference count is decremented.  This means that if there are no other
references to the parent bat, it too may be swapped out.  When the
view's physical reference count goes up to one, the parent's physical
reference count is also incremented, so that it is loaded again if it
had been swapped out.