re-added sql/src/sql/stats.sql (in Jun2010 branch only)
While sql/src/sql/stats.sql is no longer required in the default branch,
the Jun2010 branch appears not yet ready for it's removal.
At least the tests in sql/src/benchmarks/tpch/Tests/ stiil require it,
as the definition of SQL function optimizer_stats() does not seems to
happen elsewhere in the Jun2010 branch (while it does in the default).