Attempt to deal with views on varheaps.
If a BAT is a view on another BAT, that can mean several things.
- sharing BUN heap (only if no VAR heap)
- sharing BUN heap + VAR heap
- sharing VAR heap
When sharing the VAR heap, two situations were possible:
- the VAR heap structure is shared (and hence also the actual data)
- the VAR heap structure is private, but the base pointer points into
  the parent's VAR heap data.
This checkin gets rid of the latter possibility.  When the VAR heap is
shared, it is not completely shared.

A VAR heap structure is "owned" by one BAT, the parent.  Only when the
parent BAT is destroyed can the heap structure and accompanying data
be freed.  The owner of the VAR heap structure is indicated by the
parentid in the VAR heap structure.  This value must now always be the
ID of the owner BAT (and hence is always greater than zero).  If the
parentid value differs from a BAT's own ID, the var heap structure is
owned by another BAT and hence the BAT is a view on this other BAT.

Note that this checkin removes the possibility of a view's var heap
base pointer pointing somewhere halfway into the parent's var heap
data, but this situation doesn't make any sense since the BUN heap
contains offsets relative to the start of the VAR heap data, which
would be lost in this situation.

This fixes bug 2553.