Hooks for admission control The dataflow block execution has been extended with a hook for admission and wrapup. They take effect when the instruction is about to be added to or released from the queue.
The hooks context is a single stack frame.

These hooks are intended to finetune the scheduling using
more knowlegde about the context, or to implement a dynamic
dataflow crawler where the instructions to be executed are
determined by e.g. sampling first.

The stack structure is changed, so a clean rebuild may be required.