This code is incomplete. The gethostbyname sets the h_errno or errno depending on the error. But it never sets both.
if you give an unknown host,
getbostbyname("whatever");
the *h_errno* is set as *Unknown* and errno it is not set, i.e., it is *0*, therefore, strerror will print Success.

In case you give a invalid host name like:
gethostbyname(0);
the *errno* is set as *Invalid* and *h_errno* it is not set, i.e., it is *0*, therefore, hstrerror will print Success.