make test for bug 2645 work independently of degree of parallelism
As mentioned the bug report, printing the MAL plan (EXPLAIN)
of the query to check whether the desired mat.slice() call is
generated, is not convenient, as the MAL plan highly depends on
the degree of parallelism (number of CPU cores / threads)
used/available.
As althernative, we now exploit the hidden feature of collection
the execution TRACE in a table without returning it as result
(set trace = 'none';) and then query that table to check
whether mitosis was not activated (not mat.*() calls in the MAL
plan) or otherwise there is at least one mat.slice() call.

Caveat:
This might not work on Windows until bug 2612 is fixed.