Change rendering scheme to actually drop content when it doesn't fit.
This makes a change from always showing what is there to actually
dropping content (no longer displaying it).  This is done in two ways,
first columns are dropped when they would cause the result no longer to
be written on a single line.  This is indicated by the > arrow on the
right of the output at the header, and mentioned below the query result
output.
The second way is to cut strings and print them with a minimal width of
10 chars and 3 rows.  String cutting is also indicated by a trailing ...
(that extends into the border space) and a '>' border marker.

In addition, a couple of improvements have been added:
- terminal width is based on stdout, not stdin, which means that piping
  something into mclient doesn't fix its output width to 80 columns, but
  piping stdout somewhere unlimits its output width to whatever it
  needs.