removed unused (and mostly redundant) embedded libraries & servers for M4 (MIL), XQuery & M5 (MAL)
As discussed with Fabian, Sjoerd, and Niels,
to further clean-up the code base,
we remove the mostly cut-and-paste
(with minor canonical changes) replicated code
for embedded libraries & servers for M4 (MIL),
XQuery and M5 (MAL).
For now, we keep the embedded library & server for SQL.

If plain M5 (MAL) embedded library & server will be needed (again),
we should not again replicate the code, but rather move the
SQL embedded code to a top-level, say, "embedded" or "servers" directory,
that contains all shared code only once.

TODO:
Clean-up and update also the documentation related to embedded
libraries & serves (in particular on the web site) accordingly.