modpath: automatically deduce monet_mod_path when unset
Locate the modpath dirs based on the executable location.  Use platform
specific hacks to find the executable location.  The automatic
determination of the modpath can overridden by setting monet_mod_path,
such as what testing will want to do, or platforms where we can't get
the executable path.

This is not only just "nice", but also very handy for bundle installs,
or merovingians started on random directories serving as dbfarm, as it
won't require to rewrite the config file, or override the modpath in the
call invocation.

Note: get_mserver_bin_path isn't implemented for Windows, but there
appears to be a similar functionality to get the executable path.