instructionCall: backout changes from 4cef09f7b667
Not sure how the original commit protected anything.  margin is so big,
that it is bigger than most string lengths, resulting in a negative
pointer position, which on little-endian platforms apparently is
something positive, while on big-endian platforms it isn't.  This causes
the TypeException messages on big-endian platforms to be empty, causing
many tests to have different output.