configure: tune CFLAGS for regular (non-Darwin) targets more
It seems -funroll-all-loops doesn't add much in performance, hence drop
the entire flag altogether (GCC manpage discourages it anyway).
Since our performance, however, is stilll not back to the level of rev
39707, explicitly add -fomit-frame-pointer, as it used to be, since we
don't intend to debug the resulting binary (due to our mutual
exclusiveness of debug/optimize flags), we can force the flag to be in
use.

Hopefully this brings us back to the speed of 39707.