authorisation: no longer support access restrictions on scenarios
I guess this feature is hardly used (if at all), and needlessly
complicates the codebase.  Any access to the server now means access to
all available scenarios.  This has been the default since the
introduction of M5.
All old interfaces are kept equal, but scenario arguments are ignored.
On the next ABI bump for libmonetdb5, these should be removed.