mero logging: only stop logging when all buffers have been emptied
Due to race conditions, it could happen that last parts (in particular
the "Merovingian X.Y stoppped") of the logoutput wouldn't be written,
because the logger was terminated before all available content was
written.

Now only return when select indicated nothing is available for writing
to the log.  This should be safe, since all producers have been stopped
already at the point we start waiting for the logthread to end.