JDBC: rework ResultSet timestamp/time parsing
bug #2781 indicated a mismatch between what the server returned, and
what Java would print by default for time values.

The big issue here is working with timezones in the correct way.  We now
rely more on the default Java parsers to get this right.  For fields
with a timezone stored on the server, we can rely on the server doing
the work, but for fields without, we have to do zone calculations,
possibly correcting to what timezone the caller thinks the server is
using.