JDBC: rework PStimezone test to make sense
Next to making this test sort of understandable in its output, also make
the test resistant against local settings (timezone) by forcing it to
use Africa/Windhoek as default timezone.

The test is now structured as follows:
- for a timestamp, timestamp with time zone, time and time with time zone
- the same date (01-01-1970 00:00:00 UTC) is inserted
  * without any given calendar information (hence uses default Windhoek)
  * with UTC timezone
  * with America/Los_Angeles timezone
  * with arbitrary GMT+04:15 timezone (rounded to +04:00 it seems)
- query the dates and retrieve them
  * as String (aka what the server sent to us over MAPI)
  * without any given calendar information (default Windhoek)
  * with America/Los_Angeles timezone
  * with UTC timezone

For all fields with time zone, the same output regardless the timezone
should be there, since its all translated back to the local environment,
and properly maintained over zones by the server/client.
For all fields without time zone, offsets occur, but when writing and
reading is done with the same timezone, the output should be equal.